requirements.txt
asttokens==2.0.5
backcall==0.2.0
black==21.12b0
click==8.0.3
colorama==0.4.4
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.3
et-xmlfile==1.1.0
executing==0.8.2
greenlet==1.1.2
ipykernel==6.7.0
ipython==8.0.1
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.1.2
jupyter-core==4.9.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numpy==1.22.1
openpyxl==3.0.9
pandas==1.4.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
platformdirs==2.4.1
prompt-toolkit==3.0.24
psycopg2-binary==2.9.3
pure-eval==0.2.2
Pygments==2.11.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pywin32==303
pyzmq==22.3.0
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.31
stack-data==0.1.4
threadpoolctl==3.0.0
tomli==1.2.3
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
wcwidth==0.2.5
